Elevation works
Copied a whole bunch of code and got elevation to go up and down and stop when the limit switch was active.

+++ src/main/java/frc/robot/subsystems/launcherFiring/LauncherFiringSubsystem.java
@@ -30,11 +30,11 @@ public class LauncherFiringSubsystem extends SubsystemBase
 
     }
 
-    public static void ControlLaunchMotor1(int power)
+    public static void ControlLaunchMotor1(double power)
     {
         launchMotor1.set(power);
     }
-    public static void ControlLaunchMotor2(int power2)
+    public static void ControlLaunchMotor2(double power2)
     {
         launchMotor2.set(power2);
     }
